Implement initial version of loading DSL from the JSON file
