+++ tests/fixtures/danger.py
@@ -2,7 +2,7 @@ import json
 import sys
 from contextlib import contextmanager
 from io import StringIO
-from typing import Any, Dict
+from typing import Any, Dict, List, Optional
 from unittest import mock
 
 from danger_python.danger import Danger
@@ -20,3 +20,15 @@ def danger_json_input_fixture(input_json: Dict[str, Any]) -> Danger:
 
     sys.stdin = stdin
     Danger.dsl = None
+
+
+def dsl_input_fixture(modified_files: Optional[List[str]] = None) -> Dict[str, Any]:
+    return {
+        "danger": {
+            "git": {
+                "modified_files": modified_files if modified_files else [],
+                "created_files": [],
+                "deleted_files": [],
+            }
+        }
+    }
